+++ src/composables/useFlawCvssScores.ts
@@ -1,6 +1,7 @@
 import { putFlawCvssScores, postFlawCvssScores } from '@/services/FlawService';
 import { computed, ref, watch, type Ref } from 'vue';
 import type { ZodFlawType } from '@/types/zodFlaw';
+import { groupWith } from 'ramda';
 
 // TODO: This composable should be ideally refactored into a more modular
 // solution when CVSSv4 starts being used
@@ -48,9 +49,16 @@ export function useFlawCvssScores(flaw: Ref<ZodFlawType>) {
 
   const highlightedNvdCvss3String = computed(() => {
     const result = [];
-    const nvdCvssValue = nvdCvss3String.value;
-    const cvssValue = rhCvss3String.value;
+    const nvdCvssValue = flawNvdCvss3.value?.vector || '-';
+    const cvssValue = flawRhCvss3.value?.vector || '-';
     const maxLength = Math.max(nvdCvssValue.length, cvssValue.length);
+
+    if (formatScore(flawNvdCvss3.value?.score) !== formatScore(flawRhCvss3.value?.score)) {
+      result.push(
+        { char: formatScore(flawNvdCvss3.value?.score), isHighlighted: true },
+        { char: '/', isHighlighted: false }
+      );
+    }
 
     for (let i = 0; i < maxLength; i++) {
       const charFromFlaw = i < nvdCvssValue.length ? nvdCvssValue[i] : '';
@@ -61,7 +69,7 @@ export function useFlawCvssScores(flaw: Ref<ZodFlawType>) {
       });
     }
 
-    return result;
+    return groupWith((a, b) => a.isHighlighted === b.isHighlighted, result);
   });
 
   async function saveCvssScores() {